# pkg/agentos/openapi.yaml
openapi: 3.0.3
info:
  title: AgentOS API
  description: |
    AgentOS is a production-ready HTTP server for managing and running AI agents.

    ## Features
    - Session management for multi-turn conversations
    - Agent registration and execution
    - Support for tool-enabled agents
    - Memory and context management

    ## Authentication
    Currently no authentication is required. Authentication can be added via middleware.

  version: 1.0.0
  contact:
    name: Agno-Go Team
    url: https://github.com/rexleimo/agno-go
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api/v1
    description: API v1 endpoint
...
paths:
...
  /api/v1/sessions:
    post:
      tags:
        - Sessions
      summary: Create a new session
      description: Creates a new session for managing multi-turn conversations with an agent
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Sessions
      summary: List sessions
      description: Lists all sessions with optional filtering
      operationId: listSessions
      parameters:
        - name: agent_id
          in: query
          description: Filter by agent ID
          schema:
            type: string
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of sessions to return
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          description: Number of sessions to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionResponse'
                  count:
                    type: integer
                  total:
                    type: integer
  /api/v1/sessions/{id}:
    get:
      tags:
        - Sessions
      summary: Get session by ID
      description: Retrieves details of a specific session
      operationId: getSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Sessions
      summary: Update session
      description: Updates an existing session's name or metadata
      operationId: updateSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Sessions
      summary: Delete session
      description: Deletes a session and all its associated data
      operationId: deleteSession
      parameters:
        - name: id
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        '200':
          description: Session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: session deleted
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/agents/{id}/run:
    post:
      tags:
        - Agents
      summary: Run an agent
      description: Executes an agent with the provided input
      operationId: runAgent
      parameters:
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: Agent execution successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Agent execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/agents/{id}/run/stream:
    post:
      tags:
        - Agents
      summary: Stream agent run events (SSE)
      description: |
        Streams agent execution events using Server-Sent Events (SSE). Clients can optionally
        filter specific event categories (e.g. `token,reasoning,complete`) via the `types` query parameter.
      operationId: streamAgentRun
      parameters:
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
        - name: types
          in: query
          description: Comma separated list of event types to include (e.g. `run_start,token,reasoning,complete`)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: Server-sent events stream
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE payload (repeating `event:` and `data:` lines)
                example: |
                  event: run_start
                  data: {"type":"run_start","data":{"input":"Hello"}}

                  event: reasoning
                  data: {"type":"reasoning","data":{"content":"thinking...","token_count":42}}

                  event: complete
                  data: {"type":"complete","data":{"output":"Hi!", "token_count":123}}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Agent execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/teams/{id}/tools:
    get:
      tags:
        - Teams
      summary: List tools available to a team
      description: Returns aggregated tool definitions from all member agents in the team.
      operationId: listTeamTools
      parameters:
        - name: id
          in: path
          required: true
          description: Team ID
          schema:
            type: string
      responses:
        '200':
          description: Team tool definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamToolsResponse'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/knowledge/search:
    post:
      tags:
        - Knowledge
      summary: Search knowledge base
      description: Performs a vector similarity search in the knowledge base
      operationId: searchKnowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorSearchRequest'
            example:
              query: "如何使用 agno-Go 创建 Agent?"
              limit: 10
              offset: 0
      responses:
        '200':
          description: Search successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorSearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Search endpoint disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Knowledge service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/knowledge/config:
    get:
      tags:
        - Knowledge
      summary: Get knowledge base configuration
      description: Returns available chunkers, vector databases, and embedding model information
      operationId: getKnowledgeConfig
      responses:
        '200':
          description: Configuration information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeConfigResponse'
        '503':
          description: Knowledge service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/knowledge/content:
    post:
      tags:
        - Knowledge
      summary: Add content to knowledge base
      description: Ingests content into the knowledge base (chunking, embedding, storage)
      operationId: addContent
      parameters:
        - name: chunker_type
          in: query
          description: Chunker type to use (only for text/plain)
          schema:
            type: string
            enum: [character, sentence, paragraph]
            default: character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContentRequest'
          text/plain:
            schema:
              type: string
              example: "This is sample text content to be added to the knowledge base."
      responses:
        '201':
          description: Content added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddContentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Ingestion disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ingestion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Knowledge service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreateSessionRequest:
      type: object
      required:
        - agent_id
      properties:
        agent_id:
          type: string
          description: ID of the agent to associate with this session
          example: gpt-4-agent
        user_id:
          type: string
          description: Optional user ID for the session
          example: user-123
        name:
          type: string
          description: Optional name for the session
          example: Customer Support Chat
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata for the session
    UpdateSessionRequest:
      type: object
      properties:
        name:
          type: string
          description: Updated name for the session
        metadata:
          type: object
          additionalProperties: true
          description: Updated metadata for the session
    SessionResponse:
      type: object
      properties:
        session_id:
          type: string
          description: Unique session identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        agent_id:
          type: string
          description: Associated agent ID
          example: gpt-4-agent
        user_id:
          type: string
          description: Associated user ID
          example: user-123
        name:
          type: string
          description: Session name
          example: Customer Support Chat
        created_at:
          type: string
          format: date-time
          description: Timestamp when session was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when session was last updated
        metadata:
          type: object
          additionalProperties: true
          description: Session metadata
    AgentRunRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: string
          description: Input text to send to the agent
          example: What is the weather in San Francisco?
        session_id:
          type: string
          description: Optional session ID for multi-turn conversations
          example: 550e8400-e29b-41d4-a716-446655440000
        stream:
          type: boolean
          description: Deprecated. Use `/api/v1/agents/{id}/run/stream` for SSE streaming.
          default: false
    AgentRunResponse:
      type: object
      properties:
        content:
          type: string
          description: Agent's response content
          example: The weather in San Francisco is currently sunny with a temperature of 65°F.
        session_id:
          type: string
          description: Session ID if provided
          example: 550e8400-e29b-41d4-a716-446655440000
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata about the run
          properties:
            agent_id:
              type: string
            loops:
              type: integer
              description: Number of tool-calling loops executed during the run
            usage:
              $ref: '#/components/schemas/UsageMetrics'
            reasoning:
              $ref: '#/components/schemas/ReasoningSummary'
...
    TeamToolsResponse:
      type: object
      properties:
        team_id:
          type: string
          description: Team identifier
          example: "team-tools"
        tools:
          type: array
          description: Aggregated tool definitions from all member agents
          items:
            $ref: '#/components/schemas/ToolDefinition'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: agent not found
        message:
          type: string
          description: Detailed error message
          example: agent with ID 'non-existent-agent' not found
        code:
          type: string
          description: Error code
          example: AGENT_NOT_FOUND
    VectorSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query text
          example: "如何使用 agno-Go 创建 Agent?"
        limit:
          type: integer
          description: Maximum number of results to return (default 10, max 100)
          default: 10
          minimum: 1
          maximum: 100
        offset:
          type: integer
          description: Pagination offset
          default: 0
          minimum: 0
        filters:
          type: object
          additionalProperties: true
          description: Metadata filter conditions
          example:
            source: "documentation"
        collection_name:
          type: string
          description: Collection name (optional)
...
    VectorSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/VectorSearchResult'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    AddContentResponse:
      type: object
      properties:
        document_ids:
          type: array
          items:
            type: string
          description: List of created document IDs
          example: ["doc_1697123456_0", "doc_1697123456_1"]
        chunk_count:
          type: integer
          description: Number of chunks generated
          example: 2
        chunker_type:
          type: string
          description: Chunker type used
          example: "character"
        collection:
          type: string
          description: Collection documents were stored in
          example: "demo"
        message:
          type: string
          description: Success message
          example: "successfully ingested 2 chunks"